requirements.txt
fastapi==0.115.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
uvicorn[standard]==0.34.0
